+++ tests/_CanonicalFragment.ps1
@@ -18,7 +18,7 @@ function New-CanonicalOpmFragment {
         $app.Description = 'Reads PDFs'
         $app.IconPath = 'icons\app.png'
         $app.OperatingSys = 'Windows'
-        $app.OperatingSysArchitecture = 'x64'
+        $app.OperatingSysArchitecture = '64-bit'
 
         $pkg = New-OpmPackageObject
         $pkg.PackageId = '{1B2C3D4E-5F60-7182-93A4-B5C6D7E8F900}'
@@ -38,7 +38,7 @@ function New-CanonicalOpmFragment {
         $asg = New-OpmAssignmentObject
         $asg.MsEntraGroupId = '{2C3D4E5F-6071-8293-A4B5-C6D7E8F90011}'
         $asg.AssignmentType = 'Required'
-        $asg.InclusionType = 'Include'
+        $asg.InclusionType = 'included'
         $app.IntuneAssignments = @($asg)
 
         $app